# lib/api-zod/src/generated/types/brandUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Repaire Phone DZ - E-commerce API for phone repair equipment
 * OpenAPI spec version: 0.1.0
 */

export interface BrandUpdate {
  name?: string;
  description?: string;
  logoUrl?: string;
  sortOrder?: number;
}