# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Repaire Phone DZ - E-commerce API for phone repair equipment
 * OpenAPI spec version: 0.1.0
 */
import type { Address } from './address';
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  userId?: number;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userEmail?: string | null;
  status: OrderStatus;
  items?: OrderItem[];
  subtotal: number;
  discount?: number;
  /** @nullable */
  couponCode?: string | null;
  shipping?: number;
  total: number;
  shippingAddress?: Address;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}