# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Repaire Phone DZ - E-commerce API for phone repair equipment
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: UserRole;
  isBlocked?: boolean;
  totalOrders?: number;
  totalSpent?: number;
  createdAt: string;
}